fix: Label dtype

+++ src/model.py
@@ -37,7 +37,7 @@ def train_model():
     # Convert the data to a HuggingFace dataset
     #labels = df[['title_or_tasks', 'requirements', 'bad']].values
     dataset = Dataset.from_dict(dict(text=df.cleaned_text.tolist(),
-                                     label=df.bad.tolist()))
+                                     label=df.bad.astype(float).tolist()))
 
     # Split the dataset into training and validation sets
     splits = dataset.train_test_split(train_size=0.8)
